requirements.txt
asgiref==3.4.1
Django==3.2.6
django-dotenv==1.4.2
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.2
psycopg2-binary>=2.9.1
pytz==2021.1
sqlparse==0.4.1